pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "soda-mmqc"
version = "0.1.0"
description = "SODA MMQC - A tool for evaluating figure caption quality"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "Pillow>=10.0.0",
    "nltk>=3.8.1",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "tqdm",
    "pytest>=7.0.0",
    "plotly>=5.13.0",
    "pandas>=1.3.0",
    "streamlit>=1.31.0",
    "accelerate>=0.27.0",
    "watchdog>=3.0.0",
    "tenacity>=8.2.0"
]

[project.scripts]
run = "soda_mmqc.scripts.run:main"
curate = "soda_mmqc.scripts.curate:main"
init = "soda_mmqc.scripts.run:initialize_main"
viz = "soda_mmqc.scripts.visualize:main"

[tool.hatch.build.targets.wheel]
packages = ["soda_mmqc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v" 